+++ CHANGELOG.md
@@ -4,8 +4,11 @@ All notable changes to easymde will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-<!--## [Unreleased]-->
-Merge 2.7.0 with https://github.com/Ionaru/easy-markdown-editor/pull/71 @jeroenvo
+## [Unreleased/Forked]
+- Merge 2.7.0 with https://github.com/Ionaru/easy-markdown-editor/pull/71 @jeroenvo
+- replace MouseEvent(click) with .click() for IE compat
+- Fix status bar update when dragging but not dropping
+- 
 
 ## [2.7.0] - 2019-07-13
 ### Added